requirements-soft.txt
streamdeck
pillow
hidapi
pynput
nuitka
pystray